refactor: change some things from PR
