+++ frontend/public/i18n/merge-game/en.json
@@ -11,7 +11,7 @@
   "gameSetup": {
     "title": "Game Setup",
     "subtitle": "Test your logic and associative thinking",
-    "description1": "Find hidden connections and group the words into sets of 4 that share the same idea. After a correct match, a category and related questions will appear.",
+    "description1": "Find hidden connections and group the words into sets of 4 in row that share the same idea. After a correct match, a category and related questions will appear.",
     "description2": "In the “Theory” section you can view the categories, questions, and correct answers.",
 
     "mode": {
