+++ frontend/public/i18n/merge-game/ru.json
@@ -11,7 +11,7 @@
   "gameSetup": {
     "title": "Настройка игры",
     "subtitle": "Проверь свою логику и ассоциативное мышление",
-    "description1": "Найди скрытые связи и объедини слова в группы по 4 с общей идеей. После правильной сборки появится категория и вопросы по теме.",
+    "description1": "Найди скрытые связи и объедини слова в группы по 4 в ряд с общей идеей. После правильной сборки появится категория и вопросы по теме.",
     "description2": "В разделе «Теория» можно посмотреть категории, вопросы и правильные ответы.",
 
     "mode": {
