+++ frontend/src/app/components/games/merge-game/components/cat/cat.html
@@ -1,4 +1,4 @@
-<svg width="110" height="145" viewBox="0 0 271 145" fill="none" xmlns="http://www.w3.org/2000/svg">
+<svg width="100" height="auto" viewBox="0 0 271 145" fill="none" xmlns="http://www.w3.org/2000/svg">
   <path
     d="M237.938 113.745C239.029 111.858 241.206 106.717 241.392 100.843C241.626 93.441 236.012 83.898 236.127 79.9606C236.241 76.0233 236.316 68.4652 228.18 58.6346C227.542 43.3857 223.854 11.1301 214.205 4.09911C208.808 0.118643 187.874 29.6307 187.623 35.4589C181.318 34.7926 167.426 34.3558 162.301 37.9382C153.871 25.2244 135.252 0.256727 128.219 2.09653C119.427 4.39629 112.055 55.073 117.125 70.1826C115.11 76.1857 112.763 86.9886 113.329 97.1969C99.2664 95.0923 70.1234 97.4995 53.2359 119.683C50.2633 117.279 44.9567 109.231 47.9154 96.6932C51.6139 81.0211 46.119 58.245 37.3265 50.8652C28.5341 43.4854 19.1861 38.0765 9.9442 42.4279C0.702308 46.7793 -0.240941 55.9135 5.89197 57.9762C12.0249 60.039 27.7398 60.0852 28.8684 75.6301C29.997 91.175 31.3051 98.4032 29.1674 103.15C27.0298 107.896 30.351 128.8 38.1899 130.604L46.7988 130.511L95.9806 129.976C98.216 134.989 103.997 142.374 114.314 142.373C128.444 142.371 131.076 136.776 131.17 129.594L225.319 128.571L225.37 128.57C226.625 132.78 230.887 138.783 241.178 139.105C254.171 139.51 256.844 132.659 257.05 128.226C257.081 127.554 257.055 126.938 257 126.411C256.583 122.406 253.695 113.846 238.717 113.738C238.454 113.737 238.195 113.739 237.938 113.745Z"
     fill="#EFEFEF"
