+++ frontend/src/app/components/games/merge-game/components/quiz/quiz.scss
@@ -133,7 +133,7 @@
 
   &__cat {
     position: absolute;
-    top: -96px;
+    top: -49px;
     left: 15px;
     z-index: 3;
   }
