+++ frontend/src/app/components/games/merge-game/components/theory/theory.html
@@ -1,5 +1,5 @@
 <div class="nav">
-  <tui-textfield tuiChevron>
+  <tui-textfield tuiChevron [tuiTextfieldCleaner]="false">
     <input
       tuiSelect
       [ngModel]="selectedCategory()"
