+++ frontend/src/app/components/games/merge-game/components/theory/theory.scss
@@ -1,6 +1,6 @@
 :host {
   width: 100%;
-  max-width: 1200px;
+  max-width: 1500px;
   display: flex;
   flex-direction: column;
   gap: 1.5rem;
