+++ frontend/src/app/components/games/merge-game/merge-game.scss
@@ -1,9 +1,10 @@
 :host {
   width: 100vw;
+  max-width: 1500px;
   flex: 1;
   display: flex;
   flex-direction: column;
-  padding: 1rem;
+  padding: 1rem 20px;
   box-sizing: border-box;
 
   .header {
