+++ frontend/src/app/components/games/merge-game/models/ai-methods.interface.ts
