--- frontend/src/app/core/guards/settings-guard.ts
@@ -1,14 +1,0 @@
-import { inject } from '@angular/core';
-import { CanActivateFn, Router } from '@angular/router';
-import { GameService } from '../../components/games/merge-game/services/game-service';
-
-export const settingsGuard: CanActivateFn = () => {
-  const gameService = inject(GameService);
-  const router = inject(Router);
-
-  if (gameService.statusGame() === 'playing') {
-    return router.createUrlTree(['/merge-game/board']);
-  }
-
-  return true;
-};
